Breadcrumbs updated + projects moved to /data and minor fixes
