+++ src/utils/text-utils.ts
@@ -1,0 +1,3 @@
+export const capitalize = (s: string) => {
+  return s.charAt(0).toUpperCase() + s.slice(1);
+};
